scripts/flip/pyproject.toml
[project]
name = "flip"
version = "0.1.0"
description = "Image file cleaner - strips metadata and renames with random numbers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "pillow>=12.0.0",
]

[project.scripts]
flip = "main:main"
